Added a new lua script for 1024x768
Also made the login script load the right one based on the current
resolution, and made the current resolution available to the scripts.
Also changed AddScaleFactorToButton() to take floats instead of ints.

+++ TSOClient/TSOClient/LUI/UIScreen.cs
@@ -377,7 +377,7 @@ namespace TSOClient.LUI
         /// <param name="StrID">The string ID of the button.</param>
         /// <param name="ScaleX">The scalefactor on the x-axis.</param>
         /// <param name="ScaleY">The scalefactor on the y-axis.</param>
-        public void AddScalefactorToButton(string StrID, int ScaleX, int ScaleY)
+        public void AddScalefactorToButton(string StrID, float ScaleX, float ScaleY)
         {
             for (int i = 0; i < m_UIElements.Count; i++)
             {
